pyproject.toml
[project]
name = "tef-master-cloud"
version = "0.1.0"
description = "TEF Master Cloud - AI Powered French Learning"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "streamlit",
    "streamlit-option-menu",
    "google-generativeai",
    "firebase-admin",
    "ollama",
    "duckduckgo-search>=5.0",
    "beautifulsoup4"
]

[tool.uv]
dev-dependencies = []
